Move the support methods for image installations
Move the support methods for CD/DVD and ISO image installations to
the subpackage for payload, so it is clear that it is only related
to the payload.
