+++ pyanaconda/payload/image.py
